Recomm. are no longer done on fixed intervals but after transcript
Recommendations are no longer done on a fixed time basis, but are scheduled to
be executed after a new text segment is passed to the recomm. engine.

The `summaryAPI.recoInterval` property is replaced by a `recoDelay` property.
This property allows to add a delay between provinding the new transcript and
making a recomm. request (to allow for the recomm. engine to process the new
input).